requirement.txt
python-decouple
djangorestframework
django-cors-headers